% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_layers.R
\name{add_layers}
\alias{add_layers}
\title{Add ggplot layers to a Plinko board}
\usage{
add_layers(board, ...)
}
\arguments{
\item{board}{A \code{\link[=plinko_board]{plinko_board()}} object}

\item{...}{Objects that can be added to a ggplot object using \code{+}.}
}
\description{
Add ggplot layers to a Plinko board.
}
