% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plinko_board.R
\name{create_slots}
\alias{create_slots}
\title{Create slots the balls will fall into}
\usage{
create_slots(board)
}
\arguments{
\item{x}{A plinko_board}
}
\description{
Create slots the balls will fall into
}
