% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_frames.R
\name{filter_frames}
\alias{filter_frames}
\title{Filter frames in a Plinko board}
\usage{
filter_frames(board, ...)
}
\arguments{
\item{board}{A \code{\link[=plinko_board]{plinko_board()}} object}

\item{...}{A filtering expression (like one passed to \code{filter()}) that uses
columns of \code{frames(board)} to select the frames to keep.}
}
\description{
Filter the frames in Plinko board. Useful for testing animations without
needing to render the full animation.
}
