% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_coord.R
\name{modify_coord}
\alias{modify_coord}
\title{Modify coordinate system used to display a Plinko board}
\usage{
modify_coord(board, name, mapping = aes(), ...)
}
\arguments{
\item{board}{A \code{\link[=plinko_board]{plinko_board()}} object}

\item{...}{Arguments to \code{coord_fixed()} you wish to change.}
}
\description{
Modify the coordinate system used to display a \code{\link[=plinko_board]{plinko_board()}}.
\code{\link[=plinko_board]{plinko_board()}} objects use a \code{coord_fixed()} coordinate system
with several default parameters based on the properties of the board;
this function allows you to override those parameters.
}
