% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_layer.R
\name{modify_layer}
\alias{modify_layer}
\title{Modify layers/geoms used to display a Plinko board}
\usage{
modify_layer(board, name, mapping = NULL, ...)
}
\arguments{
\item{board}{A \code{\link[=plinko_board]{plinko_board()}} object}

\item{name}{Name corresponding to the geom to modify: one of \code{"slot_edges"},
\code{"pins"}, \code{"balls"}, \code{"paths"}, or \code{"dist"}.}

\item{mapping}{An \code{aes()} object containing entries in the aesthetic mapping
for this geom that you wish to change. Missing entries are left as-is.}

\item{...}{Arguments to the geom you wish to change.}
}
\description{
Modify the geoms on layers used to display the slots, pins, paths, balls,
or distribution in a \code{\link[=plinko_board]{plinko_board()}}. \code{\link[=plinko_board]{plinko_board()}} objects use several
geoms to produce the plot of the board; this function allows you to modify
arguments and aesthetic mappings passed to those geometries.
}
