% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plinko_board.R
\name{plinko_board}
\alias{plinko_board}
\alias{plinko_board.numeric}
\alias{plinko_board.distribution}
\title{Construct a Plinko board}
\usage{
plinko_board(x, ...)

\method{plinko_board}{numeric}(
  x,
  n_bin,
  bin_width,
  n_ball = NULL,
  center = NULL,
  limits = NULL,
  row_ratio = 2,
  frames_till_drop = 4,
  slot_height = NULL,
  stack = TRUE,
  ...
)

\method{plinko_board}{distribution}(
  x,
  n_bin = NULL,
  bin_width = NULL,
  n_ball = 50,
  sampling = c("quantiles", "random"),
  center = NULL,
  ...
)
}
\arguments{
\item{x}{The final values that the Plinko board will "sediment" to.}

\item{...}{Arguments passed on to lower-level implementations of \code{plinko_board()}.
Currently all implementations call down to \code{plinko_board.numeric()}, so \code{...}
may include arguments to that function not otherwise defined.}

\item{n_bin}{The number of bins used in the Binomial approximation of \code{x}.}

\item{bin_width}{The width of the bins used in the Binomial approximation of \code{x}}

\item{n_ball}{The number of balls to drop. Defaults to \code{length(x)}.}

\item{center}{The location at which to drop balls. Defaults to \code{mean(x)}.}

\item{limits}{The minimum and maximum value shown on the board. Defaults to
\code{range(x)}.}

\item{row_ratio}{The ratio between the width of the slots and the height of
a single row. Defaults to \code{2}}

\item{frames_till_drop}{The number of frames in between each drop of a ball.
Defaults to \code{4}.}

\item{slot_height}{Height of the slots. If \code{NULL} (the default), determined
automatically based on the height of the highest bin plus a bit of
overhead.}

\item{stack}{Should the balls stack on top of each other? Default \code{TRUE}.}

\item{sampling}{For \code{plinko_board.distribution}, how to "sample" balls from the
distribution. Use \code{"quantiles"}, uses quantiles of the distribution; if
\code{"random"}, samples balls randomly.}
}
\value{
An object of class \code{c("plinko_board", "list")}.
}
\description{
Constructs a semi-deterministic Plinko board. The board is semi-deterministic
in that the final values are predetermined, but the paths to get from the
drop point to the values are random.
}
