% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pres_pred_2020}
\alias{pres_pred_2020}
\title{Probabilistic forecast of the 2020 US Presidential election from 538}
\format{
A data frame with 539 rows and 11 variables:
\describe{
\item{cycle}{The election cycle: \code{2020}}
\item{branch}{The kind of race this forecast pertains to: \code{"President"}}
\item{model}{The model type: \code{"polls-plus"}}
\item{modeldate}{Date of the model run, as a \link{Date} object: \code{"2020-10-16"}}
\item{candidate_inc}{Name of the incumbent: \code{"Trump"}}
\item{candidate_chal}{Name of the challenger: \code{"Biden"}}
\item{evprob_inc}{Chance that the incumbent wins \code{total_ev} electoral votes: \link{numeric} between \code{0} and \code{1}}
\item{evprob_chal}{Chance that the challenger wins \code{total_ev} electoral votes: \link{numeric} between \code{0} and \code{1}}
\item{total_ev}{Number of electoral votes in question: \link{integer} between \code{0} and \code{538}}
\item{timestamp}{Date and time the simulations were run, as a \link{POSIXct} object: \code{"2020-10-16 21:08:11 EST"}}
\item{simulations}{Number of simulations run: \code{40000}}
}
}
\source{
This dataset contains the electoral vote output of
\href{https://projects.fivethirtyeight.com/2020-election-forecast/}{FiveThirtyEight's 2020 US Presidential Election Forecast}
as of Oct 16, 2020. The original dataset is licensed under the
\href{https://creativecommons.org/licenses/by/4.0/}{CC BY 4.0 license}. The dataset
was obtained from the
\href{https://projects.fivethirtyeight.com/2020-general-data/presidential_ev_probabilities_2020.csv}{FiveThirtyEight website},
and its metadata description is based on the dataset description in the
\href{https://github.com/fivethirtyeight/data/tree/master/election-forecasts-2020}{FiveThirtyEight Github repository}.
The dataset has been only been slightly modified: date formats were converted
from strings into R date objects; empty columns related to third party candidates
were dropped; and the data frame was sorted by \code{total_ev}.
}
\description{
A dataset containing the predicted probabilities for each major party candidate
to receive different numbers of electoral votes in the 2020 US Presidential
election according to FiveThirtyEight as of October 16, 2020.
}
\examples{

library(ggplot2)

data(pres_pred_2020, package = "plinko")

ggplot(pres_pred_2020, aes(x = total_ev, y = evprob_chal)) +
  geom_col() +
  geom_vline(xintercept = 269) +
  labs(
    x = "Electoral votes for Biden",
    y = "Predicted probability",
    title = "Predictive distribution for Biden's electoral votes",
    subtitle = "FiveThirtyEight's model of the 2020 US election as of Oct 16, 2020"
  )

}
\keyword{datasets}
