% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate.plinko_board.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{animate}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{gganimate}{\code{\link[gganimate]{animate}}}
}}

