% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tween_balls.R
\name{tween_balls}
\alias{tween_balls}
\title{Tween balls in a Plinko board}
\usage{
tween_balls(board, frame_mult, ease = "bounce-out")
}
\arguments{
\item{board}{A \code{plinko_board()} object}

\item{frame_mult}{The number of frames to add per existing frame in the
animation. 4 tends to be a good number.}

\item{ease}{The easing function to use. See \code{tweenr::display_ease()} for
examples.}
}
\value{
A modifed \code{plinko_board()} with tweened frames added.
}
\description{
Adds tweening frames to the animation of ball positions in a Plinko board.
}
