% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate.plinko_board.R
\name{animate.plinko_board}
\alias{animate.plinko_board}
\title{Animate a Plinko board}
\usage{
\method{animate}{plinko_board}(
  plot,
  fps = 10,
  renderer = gganimate::gifski_renderer(),
  device = ragg::agg_png,
  start_pause = 0,
  end_pause = 0,
  width = NULL,
  height = 800,
  res = 100,
  progress = interactive(),
  show_dist = FALSE,
  ...
)
}
\arguments{
\item{plot}{A \code{\link[=plinko_board]{plinko_board()}} object}

\item{fps}{The framerate of the animation in frames/sec (default \code{10})}

\item{renderer}{The function used to render the generated frames into an
animation. Gets a vector of paths to images along with the framerate. (default \code{\link[gganimate:renderers]{gifski_renderer()}})}

\item{device}{A device function used to render images. Default \code{ragg::agg_png}
is faster and produces higher-quality output than the base \code{png} function.}

\item{start_pause, end_pause}{Number of times to repeat the first and last
frame in the animation (default is \code{0} for both)}

\item{width, height}{One of either \code{width} or \code{height} must be provided. The
other dimension is automatically determined based on the aspect ratio of
the Plinko board.}

\item{res}{The resolution of the device. This setting will govern how device
dimensions given in inches, centimeters, or millimeters will be converted
to pixels. Further, it will be used to scale text sizes and linewidths}

\item{progress}{Output progress bar and messages? Default is to only output
during interactive sessions.}

\item{show_dist}{Should an overlay of the binomial distribution for this
Plinko board be shown?}

\item{...}{Arguments passed on to the device}
}
\description{
Animate a \code{plinko_board()} object. Follows the syntax for \code{gganimate::animate()}
with a few exceptions/additions noted below.
}
