% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plinko_board.R
\name{plinko_board-properties}
\alias{plinko_board-properties}
\alias{slot_edges}
\alias{pins}
\alias{paths}
\alias{balls}
\alias{frames}
\alias{n_frames}
\title{Get properties of a Plinko board}
\usage{
slot_edges(board)

pins(board)

paths(board)

balls(board)

frames(board)

n_frames(board)
}
\arguments{
\item{board}{A \code{\link[=plinko_board]{plinko_board()}}}
}
\value{
A data frame containing the requested properties of the Plinko board.
}
\description{
Get slot edges, pins, paths, balls, or frames from a Plinko board.
}
\section{Functions}{
\itemize{
\item \code{slot_edges}: Slot edges

\item \code{pins}: Pin locations

\item \code{paths}: Ball paths from drop point to final location

\item \code{balls}: Ball final locations

\item \code{frames}: Frames in the animation giving ball locations

\item \code{n_frames}: Number of frames in the Plinko board.
}}

